requirements.txt
click>=8.0.0
Flask>=2.0.0
PyYAML>=6.0
flask-swagger-ui>=4.0.0
requests>=2.25.0
psutil>=5.8.0